fixed filter visualization (activations still raises an error) and added md5 checksum for input files

+++ CNN_SVM_main.py
@@ -14,6 +14,7 @@ from sklearn import svm
 import time
 import random
 import xml.etree.ElementTree as ET
+import hashlib
 
 from caffe.io import array_to_blobproto
 from collections import defaultdict
@@ -23,81 +24,86 @@ from compute_mean import compute_mean
 
 
 
-def vis_square(data):
-    """Take an array of shape (n, height, width) or (n, height, width, 3)
-       and visualize each (height, width) thing in a grid of size approx. sqrt(n) by sqrt(n)"""
-    print data.shape
-    # normalize data for display
-    data = (data - data.min()) / (data.max() - data.min())
-        
-    print data.shape
-
-    # force the number of filters to be square
-    n = int(np.ceil(np.sqrt(data.shape[0])))
-    padding = (((0, n ** 2 - data.shape[0]),
-               (0, 1), (0, 1))                 # add some space between filters
-               + ((0, 0),) * (data.ndim - 3))  # don't pad the last dimension (if there is one)
-    data = np.pad(data, padding, mode='constant', constant_values=1)  # pad with ones (white)
-    print data.shape
-
-    # tile the filters into an image
-    data = data.reshape((n, n) + data.shape[1:]).transpose((0, 2, 1, 3) + tuple(range(4, data.ndim + 1)))
-    data = data.reshape((n * data.shape[1], n * data.shape[3]) + data.shape[4:])
-    
-    print data.shape
-
-    plt.imshow(data); plt.axis('off')
-
-	#cv2.waitKey(0)
 
 
+def vis_square(data, filename=''):
+	"""Take an array of shape (n, height, width) or (n, height, width, 3)
+	and visualize each (height, width) thing in a grid of size approx. sqrt(n) by sqrt(n)"""
+	
+	print 'The data you are going to visualize has shape ', data.shape
+	
+	# normalize data for display
+	data = (data - data.min()) / (data.max() - data.min())
 
-def splitDataset(imgSetName, trainPercentage):
+	# force the number of filters to be square
+	n = int(np.ceil(np.sqrt(data.shape[0])))
+	padding = (((0, n ** 2 - data.shape[0]),
+		(0, 1), (0, 1))                 # add some space between filters
+		+ ((0, 0),) * (data.ndim - 3))  # don't pad the last dimension (if there is one)
+	data = np.pad(data, padding, mode='constant', constant_values=1)  # pad with ones (white)
+
+	# tile the filters into an image
+	data = data.reshape((n, n) + data.shape[1:]).transpose((0, 2, 1, 3) + tuple(range(4, data.ndim + 1)))
+	data = data.reshape((n * data.shape[1], n * data.shape[3]) + data.shape[4:])
+
+	plt.axis('off')
+	plt.imshow(data)
+
+	# Save plot if filename is set
+	if filename != '':
+		plt.savefig(filename, bbox_inches='tight', pad_inches=0)
+
+	plt.show()
+
+
+def splitDataset(imgSetName, trainPercentage, partialMd5CheckSum):
 
 
 	#Open allNamesListFile, and split the data into:
 
-	imagesTrainSet = "ImagesTrainSet.txt"  
-	imagesTestSet = "ImagesTestSet.txt"
+	imagesTrainSetName = 'ImagesTrainSet_' + partialMd5CheckSum + '.txt'  
+	imagesTestSetName = 'ImagesTestSet_' + partialMd5CheckSum + '.txt'  
 	imgSet = []
 	trainSet = []
 	testSet = []
 
 	numElements = 0
 
-	with open(imgSetName, 'r') as file:
-		for sampleName in (file.read().splitlines()):
-			imgSet.append(sampleName)
-			numElements+=1
+	if not os.path.isfile(imagesTrainSetName) or not os.path.isfile(imagesTestSetName):
 
-	rangeImageIndices = range(numElements)
-	numberTrainSample = int(round(trainPercentage*numElements))   
-	trainIndices = random.sample(rangeImageIndices, numberTrainSample) 
-	countTrain = 0
-	countTest = 0
+		with open(imgSetName, 'r') as file:
+			for sampleName in (file.read().splitlines()):
+				imgSet.append(sampleName)
+				numElements+=1
 
-	for i in range(numElements):
-		if i in trainIndices:
-			trainSet.append(imgSet[i])
-			countTrain += 1
-		else:
-			testSet.append(imgSet[i])
-			countTest += 1
+		rangeImageIndices = range(numElements)
+		numberTrainSample = int(round(trainPercentage*numElements))   
+		trainIndices = random.sample(rangeImageIndices, numberTrainSample) 
+		countTrain = 0
+		countTest = 0
 
-	random.shuffle(trainSet)
-	random.shuffle(testSet)
+		for i in range(numElements):
+			if i in trainIndices:
+				trainSet.append(imgSet[i])
+				countTrain += 1
+			else:
+				testSet.append(imgSet[i])
+				countTest += 1
 
-	trainFile = open(imagesTrainSet, "w")
-	trainFile.write("\n".join(trainSet))
-	trainFile.close()
+		random.shuffle(trainSet)
+		random.shuffle(testSet)
 
-	testFile = open(imagesTestSet, "w")
-	testFile.write("\n".join(testSet))
-	testFile.close()
+		trainFile = open(imagesTrainSetName, "w")
+		trainFile.write("\n".join(trainSet))
+		trainFile.close()
+
+		testFile = open(imagesTestSetName, "w")
+		testFile.write("\n".join(testSet))
+		testFile.close()
 
     
 
-	return [imagesTrainSet, imagesTestSet]
+	return [imagesTrainSetName, imagesTestSetName]
 
 
 
@@ -187,7 +193,10 @@ def main(argv):
 	print 'image file is "', listImagesNames
 
 
-	caffe.set_mode_cpu()
+
+	interesting_labels = ['aeroplane','bird','cat','boat','horse']
+	trainDataPercentage = 0.7
+	
 
 	if os.path.isfile(model_filename):
 	    print 'Caffe model found.'
@@ -195,16 +204,27 @@ def main(argv):
 	    print 'Caffe model NOT found...'
 	    sys.exit(2)
 
-	trainDataPercentage = 0.7
-	
-	interesting_labels = ['aeroplane','bird','cat','boat','horse']
-	trainFeaturesFileName = 'trainFeatures.b'
-	testFeaturesFileName =  'testFeatures.b'
 
+	caffe.set_mode_cpu()
+
+	#CNN creation
+	net = caffe.Net(model_filename,      # defines the structure of the model
+	               weight_filename,  # contains the trained weights
+	              caffe.TEST)     # use test mode (e.g., don't perform dropout)
+
+	extractionLayerName = 'fc25'
+	if extractionLayerName not in net.blobs:
+		raise TypeError("Invalid layer name: " + extractionLayerName)
+	
+
+
+
+	md5CheckSum = hashlib.md5(open(listImagesNames,'rb').read()).hexdigest()
+	partialMd5CheckSum = md5CheckSum[:5]
+	
 
 	#Given file listImagesNames and percentage-> listImagesNamesTrain listImagesNamesTest
-	[listImagesNamesTrain, listImagesNamesTest] = splitDataset(listImagesNames, trainDataPercentage)
-
+	[listImagesNamesTrain, listImagesNamesTest] = splitDataset(listImagesNames, trainDataPercentage, partialMd5CheckSum)
 
 
 	#Load all images in data structure -> imagesTrain imagesTest
@@ -212,54 +232,55 @@ def main(argv):
 	[filesTestNames, imagesTest, labelsTest] = createSamplesDatastructures(listImagesNamesTest, interesting_labels)
 
 
-	#CNN creation
-	net = caffe.Net(model_filename,      # defines the structure of the model
-	               weight_filename,  # contains the trained weights
-	              caffe.TEST)     # use test mode (e.g., don't perform dropout)
+	trainFeaturesFileName = 'trainFeatures_' + partialMd5CheckSum + '.b'
+	testFeaturesFileName =  'testFeatures_' + partialMd5CheckSum + '.b'
+
+	if not os.path.isfile(trainFeaturesFileName) or not os.path.isfile(testFeaturesFileName):
 
 
-	extractionLayerName = 'fc25'
-	if extractionLayerName not in net.blobs:
-		raise TypeError("Invalid layer name: " + extractionLayerName)
+		transformer = caffe.io.Transformer({'data': net.blobs['data'].data.shape})
+		transformer.set_transpose('data', (2,0,1)) #move image channels to outermost dimension 
 
-	transformer = caffe.io.Transformer({'data': net.blobs['data'].data.shape})
-	transformer.set_transpose('data', (2,0,1)) #move image channels to outermost dimension 
+		#Update the sets of images by transforming them according to Transformer
+		for  index in range(len(imagesTrain)):
+			imagesTrain[index] = transformer.preprocess('data', imagesTrain[index])
 
-
-
-	#Update the sets of images by transforming them according to Transformer
-	for  index in range(len(imagesTrain)):
-		imagesTrain[index] = transformer.preprocess('data', imagesTrain[index])
-
-	for  index in range(len(imagesTest)):
-		imagesTest[index] = transformer.preprocess('data', imagesTest[index])		
+		for  index in range(len(imagesTest)):
+			imagesTest[index] = transformer.preprocess('data', imagesTest[index])		
 
 
-	#Forward the net on all images -> featureVectorsTrain featureVectorsTest
-	featureVectorsTrain = []
-	featureVectorsTest = []
-	
-	t1 = time.time()
-	featureVectorsTrain = extractFeatures(imagesTrain, net, extractionLayerName)
-	featureVectorsTest = extractFeatures(imagesTest, net, extractionLayerName)
-	print 'Features extraction took ',(time.time() - t1) ,' seconds for ', (len(imagesTrain) + len(imagesTest)), ' images'
+		#Forward the net on all images -> featureVectorsTrain featureVectorsTest
+		featureVectorsTrain = []
+		featureVectorsTest = []
+		
+		t1 = time.time()
+		featureVectorsTrain = extractFeatures(imagesTrain, net, extractionLayerName)
+		featureVectorsTest = extractFeatures(imagesTest, net, extractionLayerName)
+		print 'Features extraction took ',(time.time() - t1) ,' seconds for ', (len(imagesTrain) + len(imagesTest)), ' images'
 
-	#Dump features in 2 files -> trainFeatures testFeatures
-	with open(trainFeaturesFileName, 'wb') as trainFeaturesFile:
-		pickle.dump((filesTrainNames, featureVectorsTrain), trainFeaturesFile)
+		#Dump features in 2 files -> trainFeatures testFeatures
+		with open(trainFeaturesFileName, 'wb') as trainFeaturesFile:
+			pickle.dump((filesTrainNames, featureVectorsTrain), trainFeaturesFile)
 
-	with open(testFeaturesFileName, 'wb') as testFeaturesFile:
-		pickle.dump((filesTestNames, featureVectorsTest), testFeaturesFile)	
+		with open(testFeaturesFileName, 'wb') as testFeaturesFile:
+			pickle.dump((filesTestNames, featureVectorsTest), testFeaturesFile)	
 
 
-	#Load features from a previously dumped file
-	with open(trainFeaturesFileName, 'rb') as trainFeaturesFile:
-		(filesTrainNames, featureVectorsTrain) = pickle.load(trainFeaturesFile)
-		featureVectorsTrain = np.array(featureVectorsTrain)
 
-	with open(testFeaturesFileName, 'rb') as testFeaturesFile:
-		(filesTestNames, featureVectorsTest) = pickle.load(testFeaturesFile)
-		featureVectorsTest = np.array(featureVectorsTest)
+	else:
+
+		#Load features from a previously dumped file
+		with open(trainFeaturesFileName, 'rb') as trainFeaturesFile:
+			(filesTrainNames, featureVectorsTrain) = pickle.load(trainFeaturesFile)
+			featureVectorsTrain = np.array(featureVectorsTrain)
+
+		with open(testFeaturesFileName, 'rb') as testFeaturesFile:
+			(filesTestNames, featureVectorsTest) = pickle.load(testFeaturesFile)
+			featureVectorsTest = np.array(featureVectorsTest)
+
+
+
+
 
 
 	#Fit a SVM model on the extracted trainFeatures
@@ -277,16 +298,18 @@ def main(argv):
 		prediction = modelSVM.predict(features)
 		print 'For image ', filesTestNames[index], ' ... Predicted: ', prediction, ' TrueLabel: ', labelsTest[index]
 
-
 	print 'SVM test took ',(time.time() - t1) ,' seconds'
 
 
-	'''
-	filters = net.params['conv1'][0].data
-	vis_square(filters.transpose(0, 2, 3, 1))
 
-	vis_square(featureVectorsTest[0])
-	'''
+
+	filters = net.params['conv1'][0].data
+	vis_square(filters.transpose(0, 2, 3, 1), filename='conv1.png')
+
+
+	#activatedFeatures = net.blobs['conv1'].data[0, :36]
+	#vis_square(activatedFeatures)
+	
 
 if __name__=='__main__':	
 	main(sys.argv[1:])
